spark_icb_warehouse/requirements.txt
apache-airflow[amazon,microsoft.mssql,postgres]==2.5.2
apache-airflow-providers-amazon==8.8.0
pyspark==3.3.1
boto3==1.33.0
